+++ lib/moxl/src/Xec/Action/Message/Publish.php
@@ -21,6 +21,11 @@ class Publish extends Action
     protected $_originid = false;
     protected $_threadid = false;
 
+    // Reply
+    protected $_replyid = false;
+    protected $_replyto = false;
+    protected $_replyquotedbodylength = 0;
+
     // OMEMO
     protected $_messageOMEMO;
 
@@ -30,11 +35,15 @@ class Publish extends Action
         if ($this->_muc) {
             Muc::message($this->_to, $this->_content, $this->_html, $this->_id,
                          $this->_replace, $this->_file, $this->_attachid, [],
-                         $this->_originid, $this->_threadid, $this->_mucreceipts, $this->_messageOMEMO);
+                         $this->_originid, $this->_threadid, $this->_mucreceipts,
+                         $this->_replyid, $this->_replyto, $this->_replyquotedbodylength,
+                         $this->_messageOMEMO);
         } else {
             Message::message($this->_to, $this->_content, $this->_html, $this->_id,
                              $this->_replace, $this->_file, $this->_attachid, [],
-                             $this->_originid, $this->_threadid, $this->_messageOMEMO);
+                             $this->_originid, $this->_threadid, $this->_replyid,
+                             $this->_replyto, $this->_replyquotedbodylength,
+                             $this->_messageOMEMO);
         }
     }